pyproject.toml
[tool.poetry]
name = "ChartExtractor"
version = "0.1.0"
description = "ChartExtractor uses computer vision to convert images of paper charts to digital data."
authors = [
    "RyanDoesMath <vcz2aj@uvahealth.org>",
    "mattbeck1 <mtb2tgk@virginia.edu>",
    "hvalenty <unc6kr@virginia.edu>",
    "charbelmarche33 <cmarche@mail.umw.edu>",
]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
numpy = "^2.1.2"
pillow = "^10.4.0"
ultralytics = "^8.3.6"
ruff = "^0.6.9"
pytest = "^8.3.3"
scikit-learn = "^1.5.2"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
packages = [{ include = "ChartExtractor" }]
